[HUDI-5362] Rebase IncrementalRelation over HoodieBaseRelation
