+++ hudi-spark-datasource/hudi-spark-common/src/main/scala/org/apache/hudi/IncrementalRelation.scala
@@ -17,130 +17,44 @@
 
 package org.apache.hudi
 
-import org.apache.avro.Schema
-import org.apache.hadoop.fs.{GlobPattern, Path}
-import org.apache.hudi.HoodieBaseRelation.isSchemaEvolutionEnabledOnRead
-import org.apache.hudi.client.common.HoodieSparkEngineContext
-import org.apache.hudi.client.utils.SparkInternalSchemaConverter
-import org.apache.hudi.common.fs.FSUtils
-import org.apache.hudi.common.model.{HoodieCommitMetadata, HoodieFileFormat, HoodieRecord, HoodieReplaceCommitMetadata}
-import org.apache.hudi.common.table.timeline.{HoodieInstant, HoodieTimeline}
-import org.apache.hudi.common.table.{HoodieTableMetaClient, TableSchemaResolver}
-import org.apache.hudi.common.util.{HoodieTimer, InternalSchemaCache}
-import org.apache.hudi.config.HoodieWriteConfig
-import org.apache.hudi.exception.HoodieException
-import org.apache.hudi.internal.schema.InternalSchema
-import org.apache.hudi.internal.schema.utils.SerDeHelper
-import org.apache.hudi.table.HoodieSparkTable
-import org.apache.spark.api.java.JavaSparkContext
+import org.apache.hudi.common.table.timeline.HoodieTimeline
+import org.apache.hudi.common.table.view.HoodieTableFileSystemView
+import org.apache.hudi.common.table.HoodieTableMetaClient
+import org.apache.hudi.hadoop.utils.HoodieInputFormatUtils.getWritePartitionPaths
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.catalyst.InternalRow
-import org.apache.spark.sql.execution.datasources.parquet.HoodieParquetFileFormat
-import org.apache.spark.sql.sources.{BaseRelation, Filter, TableScan}
+import org.apache.spark.sql.catalyst.expressions.Expression
+import org.apache.spark.sql.sources.Filter
 import org.apache.spark.sql.types.StructType
-import org.apache.spark.sql.{DataFrame, Row, SQLContext}
-import org.slf4j.LoggerFactory
+import org.apache.spark.sql.SQLContext
 
-import scala.collection.JavaConversions._
-import scala.collection.mutable
+import scala.collection.JavaConverters._
 
 /**
  * Relation, that implements the Hoodie incremental view.
  *
  * Implemented for Copy_on_write storage.
- * TODO: rebase w/ HoodieBaseRelation HUDI-5362
  *
  */
 case class IncrementalRelation(override val sqlContext: SQLContext,
-                          override val optParams: Map[String, String],
-                          private val userSchema: Option[StructType],
-                          override val metaClient: HoodieTableMetaClient,
-                          private val prunedDataSchema: Option[StructType] = None)
-  extends BaseFileOnlyRelation(sqlContext, metaClient, optParams, userSchema, Seq(), prunedDataSchema)
+                               override val optParams: Map[String, String],
+                               private val userSchema: Option[StructType],
+                               override val metaClient: HoodieTableMetaClient,
+                               private val prunedDataSchema: Option[StructType] = None)
+  extends AbstractBaseFileOnlyRelation(sqlContext, metaClient, optParams, userSchema, Seq(), prunedDataSchema)
     with HoodieIncrementalRelationTrait {
 
-  private val log = LoggerFactory.getLogger(classOf[IncrementalRelation])
+  override type Relation = IncrementalRelation
 
-  val skeletonSchema: StructType = HoodieSparkUtils.getMetaSchema
-  private val basePath = metaClient.getBasePathV2
-  // TODO : Figure out a valid HoodieWriteConfig
-  private val hoodieTable = HoodieSparkTable.create(HoodieWriteConfig.newBuilder().withPath(basePath.toString).build(),
-    new HoodieSparkEngineContext(new JavaSparkContext(sqlContext.sparkContext)),
-    metaClient)
-  private val commitTimeline = hoodieTable.getMetaClient.getCommitTimeline.filterCompletedInstants()
-
-  private val useStateTransitionTime = optParams.get(DataSourceReadOptions.READ_BY_STATE_TRANSITION_TIME.key)
-    .map(_.toBoolean)
-    .getOrElse(DataSourceReadOptions.READ_BY_STATE_TRANSITION_TIME.defaultValue)
-
-  if (commitTimeline.empty()) {
-    throw new HoodieException("No instants to incrementally pull")
+  override def imbueConfigs(sqlContext: SQLContext): Unit = {
+    super.imbueConfigs(sqlContext)
+    // TODO(HUDI-3639) vectorized reader has to be disabled to make sure IncrementalRelation is working properly
+    sqlContext.sparkSession.sessionState.conf.setConfString("spark.sql.parquet.enableVectorizedReader", "false")
   }
-  if (!optParams.contains(DataSourceReadOptions.BEGIN_INSTANTTIME.key)) {
-    throw new HoodieException(s"Specify the begin instant time to pull from using " +
-      s"option ${DataSourceReadOptions.BEGIN_INSTANTTIME.key}")
-  }
+  override def updatePrunedDataSchema(prunedSchema: StructType): Relation =
+    this.copy(prunedDataSchema = Some(prunedSchema))
 
-  if (!metaClient.getTableConfig.populateMetaFields()) {
-    throw new HoodieException("Incremental queries are not supported when meta fields are disabled")
-  }
-
-  val useEndInstantSchema = optParams.getOrElse(DataSourceReadOptions.INCREMENTAL_READ_SCHEMA_USE_END_INSTANTTIME.key,
-    DataSourceReadOptions.INCREMENTAL_READ_SCHEMA_USE_END_INSTANTTIME.defaultValue).toBoolean
-
-  private val lastInstant = commitTimeline.lastInstant().get()
-
-  private val commitsTimelineToReturn = {
-    if (useStateTransitionTime) {
-      commitTimeline.findInstantsInRangeByStateTransitionTs(
-        optParams(DataSourceReadOptions.BEGIN_INSTANTTIME.key),
-        optParams.getOrElse(DataSourceReadOptions.END_INSTANTTIME.key(), lastInstant.getStateTransitionTime))
-    } else {
-      commitTimeline.findInstantsInRange(
-        optParams(DataSourceReadOptions.BEGIN_INSTANTTIME.key),
-        optParams.getOrElse(DataSourceReadOptions.END_INSTANTTIME.key(), lastInstant.getTimestamp))
-    }
-  }
-  private val commitsToReturn = commitsTimelineToReturn.getInstantsAsStream.iterator().toList
-
-  // use schema from a file produced in the end/latest instant
-
-  val (usedSchema, internalSchema) = {
-    log.info("Inferring schema..")
-    val schemaResolver = new TableSchemaResolver(metaClient)
-    val iSchema : InternalSchema = if (!isSchemaEvolutionEnabledOnRead(optParams, sqlContext.sparkSession)) {
-      InternalSchema.getEmptyInternalSchema
-    } else if (useEndInstantSchema && !commitsToReturn.isEmpty) {
-      InternalSchemaCache.searchSchemaAndCache(commitsToReturn.last.getTimestamp.toLong, metaClient, hoodieTable.getConfig.getInternalSchemaCacheEnable)
-    } else {
-      schemaResolver.getTableInternalSchemaFromCommitMetadata.orElse(null)
-    }
-
-    val tableSchema = if (useEndInstantSchema && iSchema.isEmptySchema) {
-      if (commitsToReturn.isEmpty) schemaResolver.getTableAvroSchema(false) else
-        schemaResolver.getTableAvroSchema(commitsToReturn.last, false)
-    } else {
-      schemaResolver.getTableAvroSchema(false)
-    }
-    if (tableSchema.getType == Schema.Type.NULL) {
-      // if there is only one commit in the table and is an empty commit without schema, return empty RDD here
-      (StructType(Nil), InternalSchema.getEmptyInternalSchema)
-    } else {
-      val dataSchema = AvroConversionUtils.convertAvroSchemaToStructType(tableSchema)
-      if (iSchema != null && !iSchema.isEmptySchema) {
-        // if internalSchema is ready, dataSchema will contains skeletonSchema
-        (dataSchema, iSchema)
-      } else {
-        (StructType(skeletonSchema.fields ++ dataSchema.fields), InternalSchema.getEmptyInternalSchema)
-      }
-    }
-  }
-
-  // https://github.com/apache/hudi/pull/485
-  private val extraFilters = optParams.getOrElse(DataSourceReadOptions.PUSH_DOWN_INCR_FILTERS.key,
-    DataSourceReadOptions.PUSH_DOWN_INCR_FILTERS.defaultValue).split(",").filter(!_.isEmpty)
-
-  override protected def timeline: HoodieTimeline = {
+  protected override def timeline: HoodieTimeline = {
     if (fullTableScan) {
       metaClient.getCommitsAndCompactionTimeline
     } else if (useStateTransitionTime) {
@@ -155,168 +69,30 @@ case class IncrementalRelation(override val sqlContext: SQLContext,
                                     requiredSchema: HoodieTableSchema,
                                     requestedColumns: Array[String],
                                     filters: Array[Filter]): RDD[InternalRow] = {
-    val results = extraFilters.map(sparkSession.sessionState.sqlParser.parseExpression)
-    super.composeRDD(fileSplits, tableSchema, requiredSchema, requestedColumns, filters)
-      .filter(row => results.map(_.re))
+    super.composeRDD(fileSplits, tableSchema, requiredSchema, requestedColumns,
+      filters ++ incrementalSpanRecordFilters)
   }
 
-  override def schema: StructType = usedSchema
-
-  override def buildScan(): RDD[Row] = {
-    if (usedSchema == StructType(Nil)) {
-      // if first commit in a table is an empty commit without schema, return empty RDD here
-      sqlContext.sparkContext.emptyRDD[Row]
+  protected override def collectFileSplits(partitionFilters: Seq[Expression],
+                                           dataFilters: Seq[Expression]): Seq[HoodieBaseFileSplit] = {
+    if (includedCommits.isEmpty) {
+      List()
     } else {
-      val regularFileIdToFullPath = mutable.HashMap[String, String]()
-      var metaBootstrapFileIdToFullPath = mutable.HashMap[String, String]()
-
-      // create Replaced file group
-      val replacedTimeline = commitsTimelineToReturn.getCompletedReplaceTimeline
-      val replacedFile = replacedTimeline.getInstants.flatMap { instant =>
-        val replaceMetadata = HoodieReplaceCommitMetadata.
-          fromBytes(metaClient.getActiveTimeline.getInstantDetails(instant).get, classOf[HoodieReplaceCommitMetadata])
-        replaceMetadata.getPartitionToReplaceFileIds.entrySet().flatMap { entry =>
-          entry.getValue.map { e =>
-            val fullPath = FSUtils.getPartitionPath(basePath, entry.getKey).toString
-            (e, fullPath)
-          }
-        }
-      }.toMap
-
-      for (commit <- commitsToReturn) {
-        val metadata: HoodieCommitMetadata = HoodieCommitMetadata.fromBytes(commitTimeline.getInstantDetails(commit)
-          .get, classOf[HoodieCommitMetadata])
-
-        if (HoodieTimeline.METADATA_BOOTSTRAP_INSTANT_TS == commit.getTimestamp) {
-          metaBootstrapFileIdToFullPath ++= metadata.getFileIdAndFullPaths(basePath).toMap.filterNot { case (k, v) =>
-            replacedFile.contains(k) && v.startsWith(replacedFile(k))
-          }
-        } else {
-          regularFileIdToFullPath ++= metadata.getFileIdAndFullPaths(basePath).toMap.filterNot { case (k, v) =>
-            replacedFile.contains(k) && v.startsWith(replacedFile(k))
-          }
-        }
-      }
-
-      if (metaBootstrapFileIdToFullPath.nonEmpty) {
-        // filer out meta bootstrap files that have had more commits since metadata bootstrap
-        metaBootstrapFileIdToFullPath = metaBootstrapFileIdToFullPath
-          .filterNot(fileIdFullPath => regularFileIdToFullPath.contains(fileIdFullPath._1))
-      }
-
-      val pathGlobPattern = optParams.getOrElse(
-        DataSourceReadOptions.INCR_PATH_GLOB.key,
-        DataSourceReadOptions.INCR_PATH_GLOB.defaultValue)
-      val (filteredRegularFullPaths, filteredMetaBootstrapFullPaths) = {
-        if (!pathGlobPattern.equals(DataSourceReadOptions.INCR_PATH_GLOB.defaultValue)) {
-          val globMatcher = new GlobPattern("*" + pathGlobPattern)
-          (regularFileIdToFullPath.filter(p => globMatcher.matches(p._2)).values,
-            metaBootstrapFileIdToFullPath.filter(p => globMatcher.matches(p._2)).values)
-        } else {
-          (regularFileIdToFullPath.values, metaBootstrapFileIdToFullPath.values)
-        }
-      }
-      // pass internalSchema to hadoopConf, so it can be used in executors.
-      val validCommits = metaClient
-        .getCommitsAndCompactionTimeline.filterCompletedInstants.getInstantsAsStream.toArray().map(_.asInstanceOf[HoodieInstant].getFileName).mkString(",")
-      sqlContext.sparkContext.hadoopConfiguration.set(SparkInternalSchemaConverter.HOODIE_QUERY_SCHEMA, SerDeHelper.toJson(internalSchema))
-      sqlContext.sparkContext.hadoopConfiguration.set(SparkInternalSchemaConverter.HOODIE_TABLE_PATH, metaClient.getBasePath)
-      sqlContext.sparkContext.hadoopConfiguration.set(SparkInternalSchemaConverter.HOODIE_VALID_COMMITS_LIST, validCommits)
-      val formatClassName = metaClient.getTableConfig.getBaseFileFormat match {
-        case HoodieFileFormat.PARQUET => HoodieParquetFileFormat.FILE_FORMAT_ID
-        case HoodieFileFormat.ORC => "orc"
-      }
-
-      // Fallback to full table scan if any of the following conditions matches:
-      //   1. the start commit is archived
-      //   2. the end commit is archived
-      //   3. there are files in metadata be deleted
-      val fallbackToFullTableScan = optParams.getOrElse(DataSourceReadOptions.INCREMENTAL_FALLBACK_TO_FULL_TABLE_SCAN_FOR_NON_EXISTING_FILES.key,
-        DataSourceReadOptions.INCREMENTAL_FALLBACK_TO_FULL_TABLE_SCAN_FOR_NON_EXISTING_FILES.defaultValue).toBoolean
-
-      val sOpts = optParams.filter(p => !p._1.equalsIgnoreCase("path"))
-
-      val startInstantTime = optParams(DataSourceReadOptions.BEGIN_INSTANTTIME.key)
-      val startInstantArchived = commitTimeline.isBeforeTimelineStarts(startInstantTime)
-      val endInstantTime = optParams.getOrElse(DataSourceReadOptions.END_INSTANTTIME.key(), lastInstant.getTimestamp)
-      val endInstantArchived = commitTimeline.isBeforeTimelineStarts(endInstantTime)
-
-      val scanDf = if (fallbackToFullTableScan && (startInstantArchived || endInstantArchived)) {
-        if (useStateTransitionTime) {
-          throw new HoodieException("Cannot use stateTransitionTime while enables full table scan")
-        }
-        log.info(s"Falling back to full table scan as startInstantArchived: $startInstantArchived, endInstantArchived: $endInstantArchived")
-        fullTableScanDataFrame(startInstantTime, endInstantTime)
+      val fileSlices = if (fullTableScan) {
+        listLatestFileSlices(Seq(), partitionFilters, dataFilters)
       } else {
-        if (filteredRegularFullPaths.isEmpty && filteredMetaBootstrapFullPaths.isEmpty) {
-          sqlContext.createDataFrame(sqlContext.sparkContext.emptyRDD[Row], usedSchema)
-        } else {
-          log.info("Additional Filters to be applied to incremental source are :" + extraFilters.mkString("Array(", ", ", ")"))
+        val latestCommit = includedCommits.last.getTimestamp
 
-          var df: DataFrame = sqlContext.createDataFrame(sqlContext.sparkContext.emptyRDD[Row], usedSchema)
+        val fsView = new HoodieTableFileSystemView(metaClient, timeline, affectedFilesInCommits)
 
-          var doFullTableScan = false
+        val modifiedPartitions = getWritePartitionPaths(commitsMetadata)
 
-          if (fallbackToFullTableScan) {
-            val fs = basePath.getFileSystem(sqlContext.sparkContext.hadoopConfiguration);
-            val timer = HoodieTimer.start
-
-            val allFilesToCheck = filteredMetaBootstrapFullPaths ++ filteredRegularFullPaths
-            val firstNotFoundPath = allFilesToCheck.find(path => !fs.exists(new Path(path)))
-            val timeTaken = timer.endTimer()
-            log.info("Checking if paths exists took " + timeTaken + "ms")
-
-            if (firstNotFoundPath.isDefined) {
-              doFullTableScan = true
-              log.info("Falling back to full table scan as some files cannot be found.")
-            }
-          }
-
-          if (doFullTableScan) {
-            fullTableScanDataFrame(startInstantTime, endInstantTime)
-          } else {
-            if (metaBootstrapFileIdToFullPath.nonEmpty) {
-              df = sqlContext.sparkSession.read
-                .format("hudi_v1")
-                .schema(usedSchema)
-                .option(DataSourceReadOptions.READ_PATHS.key, filteredMetaBootstrapFullPaths.mkString(","))
-                // Setting time to the END_INSTANT_TIME, to avoid pathFilter filter out files incorrectly.
-                .option(DataSourceReadOptions.TIME_TRAVEL_AS_OF_INSTANT.key(), endInstantTime)
-                .load()
-            }
-
-            if (regularFileIdToFullPath.nonEmpty) {
-              df = df.union(sqlContext.read.options(sOpts)
-                .schema(usedSchema).format(formatClassName)
-                // Setting time to the END_INSTANT_TIME, to avoid pathFilter filter out files incorrectly.
-                .option(DataSourceReadOptions.TIME_TRAVEL_AS_OF_INSTANT.key(), endInstantTime)
-                .load(filteredRegularFullPaths.toList: _*)
-                .filter(String.format("%s >= '%s'", HoodieRecord.COMMIT_TIME_METADATA_FIELD,
-                  commitsToReturn.head.getTimestamp))
-                .filter(String.format("%s <= '%s'", HoodieRecord.COMMIT_TIME_METADATA_FIELD,
-                  commitsToReturn.last.getTimestamp)))
-            }
-            df
-          }
-        }
+        modifiedPartitions.asScala.flatMap { relativePartitionPath =>
+          fsView.getLatestMergedFileSlicesBeforeOrOn(relativePartitionPath, latestCommit).iterator().asScala
+        }.toSeq
       }
 
-      extraFilters.foldLeft(scanDf)((e, f) => e.filter(f)).rdd
+      buildSplits(fileSlices)
     }
-  }
-
-  private def fullTableScanDataFrame(startInstantTime: String, endInstantTime: String): DataFrame = {
-    val hudiDF = sqlContext.read
-      .format("hudi_v1")
-      .schema(usedSchema)
-      .load(basePath.toString)
-      .filter(String.format("%s > '%s'", HoodieRecord.COMMIT_TIME_METADATA_FIELD, //Notice the > in place of >= because we are working with optParam instead of first commit > optParam
-        startInstantTime))
-      .filter(String.format("%s <= '%s'", HoodieRecord.COMMIT_TIME_METADATA_FIELD,
-        endInstantTime))
-
-    // schema enforcement does not happen in above spark.read with hudi. hence selecting explicitly w/ right column order
-    val fieldNames = usedSchema.fieldNames
-    hudiDF.select(fieldNames.head, fieldNames.tail: _*)
   }
 }
